@@ -5,8 +5,8 @@ class Message
 
   attr_accessor :name, :email, :body
 
-  validates :name, :email, :body, :presence => true
-  validates :email, :format => { :with => %r{.+@.+\..+} }, :allow_blank => true
+  validates :name, :email, :body, presence: true
+  validates :email, format: { with: /.+@.+\..+/ }, allow_blank: true
 
   def initialize(attributes = {})
     attributes.each do |name, value|
